pyproject.toml
[build-system]
requires = ["hatchling>=1.24"]
build-backend = "hatchling.build"

[project]
name = "dummy-spatialdata"
version = "0.1.0"
description = "A package for creating arbitrary spatialdata for testing purposes."
readme = "README.md"
requires-python = ">=3.13"
license = { text = "MIT" }
authors = [
  { name = "Artür Manukyan", email = "amanukyan.umms@gmail.com" }
]
dependencies = [
  "requests>=2.31",
  "pillow>=12.1.1",
  "spatialdata>=0.7.2",
  "spatialdata-plot>=0.3.2",
  "dummy-anndata>=0.0.3",
  "geopandas>=0.14",
  "shapely>=2.0.1"
]
optional-dependencies.dev = [
  "pre-commit",
  "twine>=4.0.2",
]
optional-dependencies.test = [
  "coverage",
  "pytest",
]
# https://docs.pypi.org/project_metadata/#project-urls
urls.Documentation = "https://dummy-spatialdata.readthedocs.io/"
urls.Homepage = "https://github.com/Artur-man/dummy-spatialdata"
urls.Source = "https://github.com/Artur-man/dummy-spatialdata"

[tool.hatch.envs.hatch-test]
features = [ "test" ]

[tool.hatch.build.targets.wheel]
only-include = ["src/dummy_spatialdata", "examples"]

[tool.hatch.build.targets.wheel.sources]
"src" = ""
"examples" = "dummy_spatialdata/examples"

[tool.hatch.build.targets.sdist]
include = [
  "/src",
  "/README.md",
  "/LICENSE",
  "/examples/*"
]

[tool.hatch.metadata]
allow-direct-references = true
